tgbot/pyproject.toml
[project]
name = "rutube-tg-bot"
version = "0.1.0"
description = "Rutube telegram bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.13.1",
    "aiohttp>=3.10.5",
    "pydantic-settings>=2.5.2",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.7",
]
